Refactored redux layout
