+++ client/src/components/Navbar/Navbar.js
@@ -3,7 +3,7 @@ import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { Link, withRouter } from 'react-router-dom';
 
-import { logOutUser } from '../../store/actions/authActions';
+import { logOutUser } from '../../store/features/authentication/authActions';
 import './styles.scss';
 
 const Navbar = ({ auth, logOutUser, history }) => {
