+++ client/src/pages/Login/Login.js
@@ -7,7 +7,7 @@ import { compose } from 'redux';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 
-import { loginUserWithEmail } from '../../store/actions/authActions';
+import { loginUserWithEmail } from '../../store/features/authentication/authActions';
 import { loginSchema } from './validation';
 import './styles.scss';
